Move commands to __main__
