+++ setup.cfg
@@ -53,7 +53,7 @@ where = src
 
 [options.entry_points]
 console_scripts =
-        dakara-play = dakara_player.commands.play:main
+        dakara-player = dakara_player.__main__:main
 
 [flake8]
 max-line-length = 88
