--- tests/test_main.py
@@ -1,1 +1,0 @@
-from dakara_player import __main__ as main  # noqa F401
